+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+##### 22.04.2025
+
+> Request table Bugs resolution
+> Booking calendar overview pagination fix
+
 ##### 21.04.2025
 
 > Split Stay logic and UI placements